% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli}
\alias{ecoli}
\title{Data from E. coli}
\format{A list with 2 elements: \describe{ \item{exp}{A dataframe of
  expression values (genes in rows, arrays in columns)} \item{reg}{A
  dataframe of regulations between a transcription factor (first column) and
  a target gene (second column). The numbers in the dataframe correspond to
  row numbers in the \code{exp} dataframe of expressions.} }}
\usage{
ecoli
}
\description{
This is data from Faith et al. (2007) containing a compendium of 445 gene
expression array for 4345 genes of E. coli, downloaded from Many Microbe
Microarrays Database (version 3 build 1) available at http://m3d.mssm.edu,
and a set of 3293 known regulations between 154TF and 1211 target genes
obtained from regulonDB v4 available at http://regulondb.ccg.unam.mx .
}
\details{
Faith JJ, Hayete B, Thaden JT, Mogno I, Wierzbowski J, et al.
(2007) Large-scale mapping and validation of Escherichia coli transcriptional
regulation from a compendium of expression profiles. PLoS Biol 5(1): e8.
doi:10.1371/journal.pbio.0050008.
}
\keyword{datasets}
