% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigress.R
\name{tigress}
\alias{tigress}
\title{Trustful Inference of Gene REgulation using Stability Selection}
\usage{
tigress(expdata, tflist = colnames(expdata), targetlist = colnames(expdata),
  alpha = 0.2, nstepsLARS = 5, nsplit = 100, normalizeexp = TRUE,
  scoring = "area", allsteps = TRUE, verb = FALSE, usemulticore = FALSE)
}
\arguments{
\item{expdata}{A matrix of expression. Each row is an experiment, each column
a gene. The gene names are the column names.}

\item{tflist}{The list of TFs names. The TF names should match the names in
the expression matrix (default: all genes are considered TFs).}

\item{targetlist}{The list of targets' names. The targets' names should match
the names in the expression matrix (default: all genes are considered
targets).}

\item{alpha}{The \code{alpha} parameter for randomization in stability
selection. Should be between 0 and 1 (default \code{0.2}).}

\item{nstepsLARS}{Number of LARS steps to perform in stability selection
(default \code{5}).}

\item{nsplit}{Number of sample splits to perform in stability selection
(default \code{100})}

\item{normalizeexp}{A boolean indicating whether we should mean center and
scale to unit variance the expression data for each gene (default
\code{TRUE}).}

\item{scoring}{The method for scoring a feature in stability selection. If
\code{"area"}, the score is the area under the stability curve up to
nstepsLARS steps, as proposed by Haury et al. If \code{"max"}, the score is
the maximum value of the stability curve, as proposed by Meinshausen and
Bühlmann in the original paper (default \code{"area"}).}

\item{allsteps}{A boolean indicating whether we should output the solutions
for all values of LARS steps up to nstepsLARS, or only for nstepsLARS. It
does not cost more computation to compute all solutions (default
\code{TRUE}).}

\item{verb}{A boolean indicating verbose mode. If \code{TRUE}, print messages
about what we are doing, otherwise remain silent (default \code{FALSE}).}

\item{usemulticore}{A boolean indicating whether multicore parallel computing
is used. This requires the package \code{parallel} (default \code{FALSE}).}
}
\value{
A list of matrices (or a single matrix if \code{allsteps=FALSE}) with
  the scores of each TF x target candidate interaction. Each row corresponds
  to a TF, each column to a target.
}
\description{
This function implements the TIGRESS model of Haury et al. for gene
regulatory network (GRN) inference from gene expression data. Given a matrix
of expression data, TIGRESS infers regulation from transcription factor (TFs)
to any target gene by performing a regression from the TF expression data to
the target gene expression, and scoring the candidate TFs by a stability
selection (SS) score.
}
